Membuat server HTML & JSON statis
