+++ Challenge-3/01. Coba Server Statis/serverHTML.js
@@ -1,0 +1,20 @@
+// Membuat Server HTML Statis
+
+const http = require("http");
+const fs = require("fs");
+
+function onRequest(request, response) {
+  response.writeHead(200, { "Content-Type": "text/html" });
+  fs.readFile("./index.html", null, (error, data) => {
+    if (error) {
+      response.writeHead(404);
+      response.write("file not found");
+    } else {
+      response.write(data);
+    }
+    response.end();
+  });
+}
+
+http.createServer(onRequest).listen(8000);
+console.log("Server HTML Statis listening on 8000..");
